Change resolver interface into a single generic resolver.
